# Slightly optimise human model rendering