--- a/src/Graphics_N64.c
+++ b/src/Graphics_N64.c
@@ -9,8 +9,6 @@
 #include <GL/gl_integration.h>
 #include <GL/gl.h>
 
-/* Current format and size of vertices */
-static int gfx_stride, gfx_format;
 typedef void (*GL_SetupVBFunc)(void);
 typedef void (*GL_SetupVBRangeFunc)(int startVertex);
 static GL_SetupVBFunc gfx_setupVBFunc;
@@ -317,11 +315,6 @@ void* Gfx_LockDynamicVb(GfxResourceID vb, VertexFormat fmt, int count) {
 
 void Gfx_UnlockDynamicVb(GfxResourceID vb) { 
 	gfx_vertices = vb;
-}
-
-void Gfx_SetDynamicVbData(GfxResourceID vb, void* vertices, int vCount) {
-	gfx_vertices = vb;
-	Mem_Copy(vb, vertices, vCount * gfx_stride);
 }
 
 void Gfx_DeleteDynamicVb(GfxResourceID* vb) { Gfx_DeleteVb(vb); }