--- a/src/Graphics_Xbox.c
+++ b/src/Graphics_Xbox.c
@@ -18,10 +18,6 @@
 // A lot of figuring out which GPU registers to use came from:
 // - comparing against pbgl and pbkit
 
-// Current format and size of vertices
-static int gfx_stride, gfx_format = -1;
-
-
 static void LoadVertexShader(uint32_t* program, int programSize) {
 	uint32_t* p;
 	
@@ -460,11 +456,6 @@ void* Gfx_LockDynamicVb(GfxResourceID vb, VertexFormat fmt, int count) {
 }
 
 void Gfx_UnlockDynamicVb(GfxResourceID vb) { Gfx_BindVb(vb); }
-
-void Gfx_SetDynamicVbData(GfxResourceID vb, void* vertices, int vCount) {
-	Mem_Copy(vb, vertices, vCount * gfx_stride);
-	Gfx_BindVb(vb);
-}
 
 void Gfx_DeleteDynamicVb(GfxResourceID* vb) { Gfx_DeleteVb(vb); }