@UnknownShadow200 2023-11-09
So far it always seems to improve performance a bit in a test map with lots of entities using human model with the same skin

Some notes:
* The Direct3D11 backend needs to be investigated, as its performance sometimes seems to suffer here
* There seems to be a major pipeline stall with ClassiCube's OpenGL rendering backends with the Intel HD Graphics 2000 GPU on Windows, similar to #1047. Need to check if switching to glBufferData eliminates this problem

Nvidia GT 740 on Windows desktop:
```
Direct3D 9:   ~800 FPS to ~1010 FPS
Direct3D 11:  ~960 FPS to ~1110 FPS
ModernGL:     ~710 FPS to ~1110 FPS
OpenGL:       ~690 FPS to ~1070 FPS
```
Intel HD Graphics 2000 on Windows desktop:
```
Direct3D 9:   ~390 FPS to ~420 FPS
Direct3D 11:   ~31 FPS to  ~63 FPS 
ModernGL:      ~26 FPS to  ~42 FPS 
OpenGL:        ~26 FPS to  ~42 FPS 
Webclient:     113 FPS to  117 FPS
```
Intel UHD Graphics 600 on Windows laptop
```
Direct3D 9:   ~260 FPS to ~290 FPS
Direct3D 11:  ~210 FPS to ~220 FPS
ModernGL:     ~130 FPS to ~180 FPS
OpenGL:       ~170 FPS to ~210 FPS
```