@UnknownShadow200 2023-11-10
I performed some additional testing - and found one case where performance actually **worsened** instead of either minorly or majorly improving (and I tested that multiple times to confirm that the performance really did worsen)
Since it's old AMD Windows OpenGL drivers, which as far as I know weren't great, I am prepared to accept the small loss there (And Direct 3D9 and Direct3D 11 backends performance improved on that same machine)

Additional machines tested on:

AMD Radeon R2 on Windows laptop
```
Direct3D 9:   ~78 FPS to ~90 FPS
Direct3D 11: ~120 FPS to ~135 FPS
ModernGL:     ~47 FPS to  ~43 FPS (worsened)
OpenGL:       ~50 FPS to  ~46 FPS (worsened)
```
NVidia GeForce 320M on mac mini
```
ModernGL:    ~117 FPS to ~160 FPS
OpenGL:      ~115 FPS to ~160 FPS
```
PowerVR GE8300 on Android
```
ModernGL:     ~4 FPS to ~9 FPS
```

Based on the performance observations collected, it's quite apparent there more work needs to be put into avoiding CPU-GPU sync stalling - as its performance impact ranges can be utterly catastrophic with some devices and/or drivers.